pyproject.toml
[tool.poetry]
name = "orders-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "order_app"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
django-rest-framework = "^0.1.0"
flake8 = "^7.1.1"
django-filter = "^24.3"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
